Wrap package_graph_for_target in package_graph_trait compatible layer
This takes the new cyclic graph work and wraps it in the
package_graph_trait. There's some conversion work required as the
existing API uses Strings where we might have a PackageIdent, and
package information is represented as a originsrv::OriginPackage and
needs to be converted to the internal representation.

Signed-off-by: Mark Anderson <mark@chef.io>
Co-authored-by: Scott Macfarlane <smacfarlane@chef.io>
Co-authored-by: Mark Anderson <mark@chef.io>