Add datastore tests. The intent is to get some basic test coverage before refactoring it to use Diesel.
This revives a macro 'datastore_test' which sets up and initializes a
test database from a master template file. That probably could be
simplified/improved, but probably not worth it.

We use that macro to write some basic test coverage fro the
data_store.rs file, prior to refactoring things.

We're missing good coverage on timestamps, which are largely handled
by stored procedures.

Co-Authored-by: Ian Henry <ianjhenry00@gmail.com>
Co-Authored-by: Scott Macfarlane <smacfarlane@chef.io>
Signed-off-by: Mark Anderson <mark@chef.io>